+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-views/analyses/analysis-form-models/georeference-admin-region-model.js
@@ -1,5 +1,3 @@
-var _ = require('underscore');
-
 module.exports = {
   type: 'georeference-admin-region',
   label: _t('editor.layers.analysis-form.georeference-admin-region'),
@@ -20,14 +18,11 @@ module.exports = {
   parametersDataFields: 'admin_region_column,country',
 
   parse: function (nodeAttrs) {
-    var opts = {
+    return {
+      type: 'georeference-admin-region',
       admin_region_column: nodeAttrs.admin_region_column,
       country: nodeAttrs.country || nodeAttrs.country_column
     };
-
-    return _.extend({
-      type: 'georeference-admin-region'
-    }, opts);
   },
 
   formatAttrs: function (formAttrs, columnOptions) {
