+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-views/analyses/analysis-form-models/georeference-city-model.js
@@ -1,17 +1,13 @@
 var _ = require('underscore');
 
 module.exports = {
-  type: 'georeference-street-address',
-  label: _t('editor.layers.analysis-form.georeference-street-address'),
+  type: 'georeference-city',
+  label: _t('editor.layers.analysis-form.georeference-city'),
   fieldAttrs: {
-    street_address_column: {
+    city_column: {
       showSearch: true
     },
-    city: {
-      showSearch: true,
-      allowFreeTextInput: true
-    },
-    state: {
+    admin_region: {
       showSearch: true,
       allowFreeTextInput: true
     },
@@ -22,26 +18,22 @@ module.exports = {
   },
 
   validatorFor: {
-    street_address_column: ['required']
+    city_column: ['required']
   },
 
-  parametersDataFields: 'street_address_column,city,state,country',
+  parametersDataFields: 'city_column,admin_region,country',
 
   parse: function (nodeAttrs) {
-    var opts = {
-      street_address_column: nodeAttrs.street_address_column,
-      city: nodeAttrs.city || nodeAttrs.city_column,
-      state: nodeAttrs.state || nodeAttrs.state_column,
+    return {
+      type: 'georeference-city',
+      city_column: nodeAttrs.city_column,
+      admin_region: nodeAttrs.admin_region || nodeAttrs.admin_region_column,
       country: nodeAttrs.country || nodeAttrs.country_column
     };
-
-    return _.extend({
-      type: 'georeference-street-address'
-    }, opts);
   },
 
   formatAttrs: function (formAttrs, columnOptions) {
-    _.each(['city', 'state', 'country'], function (attr) {
+    _.each(['country', 'admin_region'], function (attr) {
       if (formAttrs[attr] && columnOptions.findColumn(formAttrs[attr], 'string')) {
         formAttrs[attr + '_column'] = formAttrs[attr];
         delete formAttrs[attr];
