+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-views/analyses/analysis-form-models/georeference-country-model.js
@@ -1,5 +1,3 @@
-var _ = require('underscore');
-
 module.exports = {
   type: 'georeference-country',
   label: _t('editor.layers.analysis-form.georeference-country'),
@@ -16,13 +14,10 @@ module.exports = {
   parametersDataFields: 'country_column',
 
   parse: function (nodeAttrs) {
-    var opts = {
+    return {
+      type: 'georeference-country',
       country_column: nodeAttrs.country_column || nodeAttrs.country
     };
-
-    return _.extend({
-      type: 'georeference-country'
-    }, opts);
   },
 
   formatAttrs: function (formAttrs) {
