+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-views/analyses/analysis-form-models/georeference-ip-address-model.js
@@ -1,5 +1,3 @@
-var _ = require('underscore');
-
 module.exports = {
   type: 'georeference-ip-address',
   label: _t('editor.layers.analysis-form.georeference-ip-address'),
@@ -16,13 +14,10 @@ module.exports = {
   parametersDataFields: 'ip_address',
 
   parse: function (nodeAttrs) {
-    var opts = {
+    return {
+      type: 'georeference-ip-address',
       ip_address: nodeAttrs.ip_address
     };
-
-    return _.extend({
-      type: 'georeference-ip-address'
-    }, opts);
   },
 
   formatAttrs: function (formAttrs) {
