+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-views/analyses/analysis-form-models/georeference-postal-code-model.js
@@ -1,5 +1,3 @@
-var _ = require('underscore');
-
 module.exports = {
   type: 'georeference-postal-code',
   label: _t('editor.layers.analysis-form.georeference-postal-code'),
@@ -20,14 +18,11 @@ module.exports = {
   parametersDataFields: 'postal_code_column,country',
 
   parse: function (nodeAttrs) {
-    var opts = {
+    return {
+      type: 'georeference-postal-code',
       postal_code_column: nodeAttrs.postal_code_column,
       country: nodeAttrs.country || nodeAttrs.country_column
     };
-
-    return _.extend({
-      type: 'georeference-postal-code'
-    }, opts);
   },
 
   formatAttrs: function (formAttrs, columnOptions) {
