+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-views/analyses/analysis-form-models/georeference-street-address-model.js
@@ -1,13 +1,17 @@
 var _ = require('underscore');
 
 module.exports = {
-  type: 'georeference-city',
-  label: _t('editor.layers.analysis-form.georeference-city'),
+  type: 'georeference-street-address',
+  label: _t('editor.layers.analysis-form.georeference-street-address'),
   fieldAttrs: {
-    city_column: {
+    street_address_column: {
       showSearch: true
     },
-    admin_region: {
+    city: {
+      showSearch: true,
+      allowFreeTextInput: true
+    },
+    state: {
       showSearch: true,
       allowFreeTextInput: true
     },
@@ -18,25 +22,23 @@ module.exports = {
   },
 
   validatorFor: {
-    city_column: ['required']
+    street_address_column: ['required']
   },
 
-  parametersDataFields: 'city_column,admin_region,country',
+  parametersDataFields: 'street_address_column,city,state,country',
 
   parse: function (nodeAttrs) {
-    var opts = {
-      city_column: nodeAttrs.city_column,
-      admin_region: nodeAttrs.admin_region || nodeAttrs.admin_region_column,
+    return {
+      type: 'georeference-street-address',
+      street_address_column: nodeAttrs.street_address_column,
+      city: nodeAttrs.city || nodeAttrs.city_column,
+      state: nodeAttrs.state || nodeAttrs.state_column,
       country: nodeAttrs.country || nodeAttrs.country_column
     };
-
-    return _.extend({
-      type: 'georeference-city'
-    }, opts);
   },
 
   formatAttrs: function (formAttrs, columnOptions) {
-    _.each(['country', 'admin_region'], function (attr) {
+    _.each(['city', 'state', 'country'], function (attr) {
       if (formAttrs[attr] && columnOptions.findColumn(formAttrs[attr], 'string')) {
         formAttrs[attr + '_column'] = formAttrs[attr];
         delete formAttrs[attr];
