+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-views/analyses/analysis-form-models/georeference-types.js
@@ -1,9 +1,9 @@
 module.exports = [
-  require('./georeference-long-lat'),
-  require('./georeference-city'),
-  require('./georeference-country'),
-  require('./georeference-admin-region'),
-  require('./georeference-postal-code'),
-  require('./georeference-ip-address'),
-  require('./georeference-street-address')
+  require('./georeference-long-lat-model'),
+  require('./georeference-city-model'),
+  require('./georeference-country-model'),
+  require('./georeference-admin-region-model'),
+  require('./georeference-postal-code-model'),
+  require('./georeference-ip-address-model'),
+  require('./georeference-street-address-model')
 ];
